LayoutTests/imported/w3c/web-platform-tests/svg/animations/pruning-first-interval-expected.txt


FAIL Timed element not active after first interval active duration changes to unresolved assert_equals: expected "rgb(0, 128, 0)" but got "rgb(255, 0, 0)"
